Simplify equality check
The current diff printout (on test failure) now fails with
> panic: cannot handle unexported field at {sysregistriesv2.V2RegistriesConf}.shortNameAliasConf:

Meanwhile, diff.ObjectGoPrintDiff is deprecated, and we are already using the
require library that allows doing the deep comparison and diff output in a single
line, so replace the manual use of reflection and diff with a one-liner.
That also allows us to drop a dependency.

Only changes tests, should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>